Get rid of all current type hints issues and partition sources
